feat. 플레이캐스트액티비티내에서 담아온 캐스트에 관한 카테고리 추가및 삭제 구현

+++ app/src/main/java/kr/dori/android/own_cast/forApiData/AuthResponse.kt
@@ -77,7 +77,7 @@ data class GetUserPlaylist(
     @SerializedName(value = "name") val name: String,
     @SerializedName(value = "imagePath") val imagePath: String,
     @SerializedName(value = "playlistId") val playlistId: Long,
-    @SerializedName(value = "totalCast") val totalCast: Long
+    @SerializedName(value = "totalCast") val totalCast: Int
 
 )
